requirements.txt
ruff
jupyterlab
fastapi[standard]
pandas
python-dotenv
psycopg2
